src/core/model_partitioner.py
from typing import List, Dict, Any, Tuple
import torch
import torch.nn as nn
import numpy as np

class ModelPartitioner:
    def __init__(self, model: nn.Module):
        """Initialize the model partitioner.
        
        Args:
            model: PyTorch model to be partitioned
        """
        self.model = model
        self.layers = self._analyze_model_layers()
        self.computation_graph = {}
        
    def _analyze_model_layers(self) -> List[Dict[str, Any]]:
        """Analyze model layers and their computational requirements."""
        layers_info = []
        total_params = 0
        
        for name, module in self.model.named_modules():
            if len(list(module.children())) == 0:  # Only leaf modules
                params = sum(p.numel() for p in module.parameters())
                total_params += params
                
                layer_info = {
                    'name': name,
                    'type': type(module).__name__,
                    'parameters': params,
                    'input_shape': None,  # Will be filled during forward pass
                    'output_shape': None,
                    'computation_cost': self._estimate_computation_cost(module)
                }
                layers_info.append(layer_info)
                
        # Normalize computation costs
        total_cost = sum(layer['computation_cost'] for layer in layers_info)
        for layer in layers_info:
            layer['relative_cost'] = layer['computation_cost'] / total_cost
            
        return layers_info
    
    def _estimate_computation_cost(self, module: nn.Module) -> float:
        """Estimate computational cost of a layer."""
        if isinstance(module, nn.Conv2d):
            # Cost = kernel_size^2 * in_channels * out_channels * output_size^2
            cost = (module.kernel_size[0] * module.kernel_size[1] * 
                   module.in_channels * module.out_channels)
        elif isinstance(module, nn.Linear):
            # Cost = input_features * output_features
            cost = module.in_features * module.out_features
        else:
            # Default cost based on parameters
            cost = sum(p.numel() for p in module.parameters())
        return float(cost)
    
    def partition_model(self, num_partitions: int, 
                       resource_weights: Dict[str, float]) -> List[List[str]]:
        """Partition the model based on computational requirements and resources."""
        # Calculate partition boundaries based on cumulative cost
        total_cost = sum(layer['computation_cost'] for layer in self.layers)
        target_cost = total_cost / num_partitions
        
        current_partition = []
        partitions = []
        current_cost = 0
        
        for layer in self.layers:
            current_partition.append(layer['name'])
            current_cost += layer['computation_cost']
            
            if current_cost >= target_cost:
                partitions.append(current_partition)
                current_partition = []
                current_cost = 0
                
        # Add remaining layers to last partition
        if current_partition:
            partitions.append(current_partition)
            
        return partitions
    
    def analyze_communication_cost(self, partitions: List[List[str]]) -> Dict[str, float]:
        """Analyze communication cost between partitions."""
        costs = {}
        for i in range(len(partitions) - 1):
            partition_a = partitions[i]
            partition_b = partitions[i + 1]
            
            # Estimate data transfer size between partitions
            transfer_size = self._estimate_transfer_size(partition_a[-1], partition_b[0])
            costs[f"partition_{i}_to_{i+1}"] = transfer_size
            
        return costs
    
    def _estimate_transfer_size(self, layer1_name: str, layer2_name: str) -> float:
        """Estimate data transfer size between layers."""
        # Implementation depends on specific model architecture
        # This is a placeholder that should be implemented based on actual model
        return 0.0
    
    def export_partitions(self, partitions: List[List[str]], export_path: str):
        """Export partitioned models to files."""
        for i, partition in enumerate(partitions):
            # Create a new model for this partition
            partition_model = self._create_partition_model(partition)
            
            # Save the partition
            save_path = f"{export_path}/partition_{i}.pt"
            torch.save(partition_model.state_dict(), save_path)
            
    def _create_partition_model(self, layer_names: List[str]) -> nn.Module:
        """Create a model from a subset of layers."""
        # Implementation depends on specific model architecture
        # This is a placeholder that should be implemented based on actual model
        return nn.Sequential()
